Исправление архитектуры
Почистил код, создал папку для отдельных элементов экранов
