+++ habits/lib/pages/screen.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 
-class HistoryPage extends StatelessWidget {
-  const HistoryPage({super.key});
+class Screen4 extends StatelessWidget {
+  const Screen4({super.key});
 
   @override
   Widget build(BuildContext context) {
     return const Scaffold(
       body: Center(
         child: Text(
-          'History',
+          'Screen4',
           style: TextStyle(fontSize: 30),
         )
       )
